// frontend/lib/types.ts
export interface EdibioUser {
    uid: string;
    email: string;
    name: string;
    phone?: string;
    photoUrl?: string;
    createdAt: string;
    trialExpiresAt?: string;
    trialClaimed?: boolean;
    subscriptionType?: 'free' | 'mobile' | 'standard' | 'premium';
    subscriptionExpiresAt?: string;
    paymentHistory?: { plan: string; amount: number; date: string; billingCycle: string; status: string; id?: string; }[];
    role?: string;
}